[REFACTOR] 공용으로 사용하는 NumberElementCollection 구현코드 중 Lotto 도메인에 필요한 코드만 추출해서 서브클래스(LottoNumberCollection)로 이동
